refactor(aio): move deployment config tests and helpers around
This commit prepares the ground for adding different types of tests.
